+++ BW/src/bw/body/io/sensors/internal/clsStomachSensor.java
@@ -26,7 +26,7 @@ public class clsStomachSensor extends clsSensorInt {
 
 	private clsBaseBody moBody;
 	
-	private float mrEnergy;
+	private double mrEnergy;
 	
 	/**
 	 * TODO (langr) - insert description 
@@ -129,7 +129,7 @@ public class clsStomachSensor extends clsSensorInt {
 	/**
 	 * @return the mrEnergy
 	 */
-	public float getEnergy() {
+	public double getEnergy() {
 		return mrEnergy;
 	}
